% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoextraplus-package.R
\docType{package}
\name{factoextraplus-package}
\alias{factoextraplus}
\alias{factoextraplus-package}
\title{factoextraplus: Extract and Visualize the Results of Multivariate Data Analyses (Extended)}
\description{
Extended version of 'factoextra' with modernized code for compatibility with current R ecosystem (ggplot2 3.5.2+, R 4.1.0+). Provides easy-to-use functions to extract and visualize the output of multivariate data analyses, including 'PCA' (Principal Component Analysis), 'CA' (Correspondence Analysis), 'MCA' (Multiple Correspondence Analysis), 'FAMD' (Factor Analysis of Mixed Data), 'MFA' (Multiple Factor Analysis) and 'HMFA' (Hierarchical Multiple Factor Analysis) functions from different R packages. It contains also functions for simplifying some clustering analysis steps and provides 'ggplot2'-based elegant data visualization.
}
\details{
factoextraplus is an extended version of factoextra with modernized code
for compatibility with current R ecosystem (ggplot2 3.5.2+, R 4.1.0+).

General resources:

\itemize{
\item \href{https://github.com/erdeyl/factoextra-modern}{factoextraplus documentation}
\item \href{https://rpkgs.datanovia.com/factoextra/}{Original factoextra documentation}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erdeyl/factoextraplus}
  \item Report bugs at \url{https://github.com/erdeyl/factoextraplus/issues}
}

}
\author{
\strong{Maintainer}: Laszlo Erdey \email{erdey.laszlo@econ.unideb.hu} (Faculty of Economics and Business, University of Debrecen, Hungary)

Authors:
\itemize{
  \item Alboukadel Kassambara \email{alboukadel.kassambara@gmail.com} (Original factoextra package author)
  \item Fabian Mundt \email{f.mundt@inventionate.de} (Original factoextra package author)
}

}
\keyword{internal}
